Replaces spa_pod_builder_add with prop and value calls
Minor change, both of these APIs do the same thing. My rational
for the change is just that `spa_pod_builder_add` seems to be
intended as a helper to implement the `spa_builder_add_object`-like
macros which can't be translated to Zig. As such, it does this odd
thing where it expects strings to be passed into varargs that are
later parsed. Since we're not using the macro, we may as well drop
all the way down to the lower level API which bypasses all that and
is easier to understand IMO anyway.